pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qqq-trading"
version = "1.0.0"
description = "QQQ intraday volatility prediction for 0DTE/1DTE straddle trading"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "lightgbm>=4.0",
    "joblib>=1.3",
    "shap>=0.43",
    "scipy>=1.11",
    "yfinance>=0.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[project.scripts]
qqq-predict = "qqq_trading.cli.predict:main"
qqq-pipeline = "qqq_trading.cli.pipeline:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
